requirements.txt
line-bot-sdk
flask
gunicorn
pymongo
imgurpython
requests
numpy
matplotlib
twstock
pandas
lxml